+++ lib/src/runner/vm/environment.dart
@@ -4,8 +4,9 @@
 
 library test.runner.vm.environment;
 
+import 'package:async/async.dart';
+
 import '../environment.dart';
-import '../../util/cancelable_future.dart';
 
 /// The environment for test suites loaded in an isolate in this Dart VM.
 class VMEnvironment implements Environment {
@@ -17,7 +18,7 @@ class VMEnvironment implements Environment {
   Uri get remoteDebuggerUrl => throw new UnsupportedError(
       "VMEnvironment.observatoryUrl is not supported.");
 
-  CancelableFuture displayPause() =>
+  CancelableOperation displayPause() =>
       throw new UnsupportedError(
           "The VM doesn't yet support Environment.displayPause.");
 }