+++ test/runner/engine_test.dart
@@ -4,7 +4,6 @@
 
 import 'package:test/src/backend/group.dart';
 import 'package:test/src/backend/state.dart';
-import 'package:test/src/backend/test.dart';
 import 'package:test/src/runner/engine.dart';
 import 'package:test/src/runner/runner_suite.dart';
 import 'package:test/src/runner/vm/environment.dart';