+++ test/runner/expanded_reporter_test.dart
@@ -16,7 +16,7 @@ void main() {
   test("reports when no tests are run", () {
     d.file("test.dart", "void main() {}").create();
 
-    var test = runTest(["test.dart"], compact: true);
+    var test = runTest(["test.dart"]);
     test.stdout.expect(consumeThrough(contains("No tests ran.")));
     test.shouldExit(0);
   });
@@ -68,7 +68,7 @@ void main() {
 }
 """).create();
 
-    var test = runTest(["--verbose-trace", "test.dart"], compact: true);
+    var test = runTest(["--verbose-trace", "test.dart"]);
     test.stdout.expect(consumeThrough(contains("dart:isolate-patch")));
     test.shouldExit(1);
   });
@@ -249,6 +249,18 @@ void main() {
           +0 ~1: skip 2
           +0 ~2: skip 3
           +0 ~3: All tests skipped.""");
+    });
+
+    test("displays a skipped group", () {
+      _expectReport("""
+          group('skip', () {
+            test('test 1', () {});
+            test('test 2', () {});
+            test('test 3', () {});
+          }, skip: true);""",
+          """
+          +0: skip
+          +0 ~1: All tests skipped.""");
     });
 
     test("runs skipped tests along with successful tests", () {